refactor: remove unnecessary single-export barrel files
- Remove barrel files with only a single export to reduce indirection
- Update imports to point directly to source files
- Keep barrel files that export from multiple subdirectories
- Simplify codebase by reducing unnecessary abstraction layers
