+++ src/features/auth/api/hooks/index.ts
@@ -1,1 +1,7 @@
+/**
+ * Barrel file for auth hooks exports
+ * 
+ * Consolidates auth hook exports for cleaner imports.
+ */
+
 export * from './useAuth';
