--- src/features/help/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useHelp';
