--- src/features/progress/components/index.ts
@@ -1,1 +1,0 @@
-export { default as ProgressWithFeedback } from './ProgressWithFeedback';
