--- src/features/progress/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useProgressFeedback } from './useProgressFeedback';
