+++ src/features/statements/index.ts
@@ -1,3 +1,8 @@
-// Barrel file for statements feature exports
+/**
+ * Barrel file for statements feature
+ * 
+ * Provides unified access to statements-related functionality
+ * including context, providers, hooks, and components.
+ */
 export { useEntries } from './hooks/useEntries';
 export { EntriesProvider } from './context/EntriesProvider';
