--- src/features/wizard/index.ts
@@ -1,2 +1,0 @@
-// Barrel file for wizard feature exports
-export { default as StatementWizard } from './components/StatementWizard';
